TV: Move PodcastsViewModel to its own file
