+++ Pocket Casts TV App/UI/Podcasts/PodcastsView.swift
@@ -1,33 +1,4 @@
 import SwiftUI
-import Combine
-
-@Observable
-class PodcastsViewModel {
-
-    private var cancellable: AnyCancellable?
-
-    enum State: Equatable, Hashable {
-        case loading
-        case ready
-        case empty
-    }
-
-    var state: State = .loading
-
-    var podcasts: [MockPodcast] = MockData.makePodcasts()
-
-    func load() {
-        //Mock data load
-        cancellable = Timer.publish(every: 1.0, on: .main, in: .common, options: nil)
-            .autoconnect()
-            .sink { [weak self] _ in
-                guard let self else { return }
-                state = .ready
-                cancellable?.cancel()
-                cancellable = nil
-            }
-    }
-}
 
 struct PodcastsView: View {
     @Environment(AppCoordinator.self) var coordinator
